// src/lib/auth/api/api.ts
export interface V1AdminEventLogsGet200Response {
    /**
     * 
     * @type {string}
     * @memberof V1AdminEventLogsGet200Response
     */
    'code': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminEventLogsGet200Response
     */
    'message': string | null;
    /**
     * 
     * @type {V1AdminEventLogsGet200ResponseData}
     * @memberof V1AdminEventLogsGet200Response
     */
    'data': V1AdminEventLogsGet200ResponseData;
}
/**
 * 
 * @export
 * @interface V1AdminEventLogsGet200ResponseData
 */
export interface V1AdminEventLogsGet200ResponseData {
    /**
     * 
     * @type {Array<V1AdminEventLogsGet200ResponseDataDataInner>}
     * @memberof V1AdminEventLogsGet200ResponseData
     */
    'data': Array<V1AdminEventLogsGet200ResponseDataDataInner>;
    /**
     * 
     * @type {number}
     * @memberof V1AdminEventLogsGet200ResponseData
     */
    'count': number;
}
/**
 * 
 * @export
 * @interface V1AdminEventLogsGet200ResponseDataDataInner
 */
export interface V1AdminEventLogsGet200ResponseDataDataInner {
    /**
     * 
     * @type {string}
     * @memberof V1AdminEventLogsGet200ResponseDataDataInner
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminEventLogsGet200ResponseDataDataInner
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminEventLogsGet200ResponseDataDataInner
     */
    'updatedAt': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminEventLogsGet200ResponseDataDataInner
     */
    'type': V1AdminEventLogsGet200ResponseDataDataInnerTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof V1AdminEventLogsGet200ResponseDataDataInner
     */
    'dataId'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminEventLogsGet200ResponseDataDataInner
     */
    'sessionUserId'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminEventLogsGet200ResponseDataDataInner
     */
    'data'?: string | null;
}